temp: upgrade to opp6
* fix cChannel api
* fix .ned files' notations

+++ quisp/test_utils/RNG.h
@@ -10,11 +10,11 @@ class TestRNG : public omnetpp::cRNG {
  public:
   void initialize(int seedSet, int rngId, int numRngs, int parsimProcId, int parsimNumPartitions, omnetpp::cConfiguration *cfg) override {}
   void selfTest() override {}
-  unsigned long intRandMax() override { return 10000; };
+  unsigned int intRandMax() override { return 10000; };
 
   // Random integer in [0,n), n < intRandMax()
-  unsigned long intRand(unsigned long n) override { return intValue; };
-  unsigned long intRand() override { return intValue; }
+  unsigned int intRand(unsigned int n) override { return intValue; };
+  unsigned int intRand() override { return intValue; }
 
   // Random double on the [0,1) interval
   double doubleRand() override { return doubleValue; };
